SOLV 2891: Profile is empty
